pyproject.toml
[tool.poetry]
name = "hw28"
version = "0.0.1"
description = "ya"
authors = ["den"]
license = "no"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
